--- plugins_/lib/sublime_lib/view/__init__.py
@@ -1,3 +1,0 @@
-# for importing the module
-from ._view import *  # NOQA
-from ._view import __all__  # NOQA
